requirements.txt
psycopg2==2.9.5
python-dotenv==1.0.1
requests==2.28.1